ASTRA/requirements.txt
reportlab>=3.6.0
PyYAML>=6.0
requests>=2.28.0
pandas>=1.5.0
